Refactor README and remove outdated performance tests
