+++ monitoring/browser/browser.test.ts
@@ -1,11 +1,11 @@
 import { sleep, streamTimeout } from "@helpers/client";
 import { getTime } from "@helpers/logger";
-import { playwright } from "@helpers/playwright";
 import { setupDurationTracking } from "@helpers/vitest";
 import { getInboxIds, getRandomInboxIds } from "@inboxes/utils";
 import { typeofStream } from "@workers/main";
 import { getWorkers, type Worker } from "@workers/manager";
 import { beforeAll, describe, expect, it } from "vitest";
+import { playwright } from "./playwright";
 
 const testName = "browser";
 describe(testName, () => {
